refactor(MiniWindow): separate network request codes into independent hook
